PCI: Return u16 from pci_find_ext_capability() and similar
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1993252
Upstream Status: ee8b1c478a9fbce9c64151ee561c124c4dcd66be

Conflict(s):
  Checking file include/linux/pci.h; Hunk #1 FAILED at 381.  This is
  due to RHEL's kABI.


commit ee8b1c478a9fbce9c64151ee561c124c4dcd66be
Author: Bjorn Helgaas <bhelgaas@google.com>
Date:   Fri Dec 4 15:14:07 2020 -0600

    PCI: Return u16 from pci_find_ext_capability() and similar

    PCI Express Extended Capabilities are in config space between offsets 256
    and 4K.  These offsets all fit in 16 bits.

    Change the return type of pci_find_ext_capability() and supporting
    functions from int to u16 to match the specification.  Many callers use
    "int", which is fine, but there's no need to store more than a u16.

+++ include/linux/pci.h
@@ -519,7 +519,7 @@ struct pci_dev {
 #endif
 	RH_KABI_USE(5, u16  acs_cap)
 #ifdef CONFIG_PCIEASPM
-	RH_KABI_USE(6, int  l1ss)	/* L1SS Capability pointer */
+	RH_KABI_USE(6, u16  l1ss)	/* L1SS Capability pointer */
 #endif
 #ifdef CONFIG_PCIEPORTBUS
 	RH_KABI_USE(7, struct rcec_ea  *rcec_ea) /* RCEC cached endpoint association */
@@ -1126,8 +1126,8 @@ u8 pci_find_capability(struct pci_dev *dev, int cap);
 u8 pci_find_next_capability(struct pci_dev *dev, u8 pos, int cap);
 u8 pci_find_ht_capability(struct pci_dev *dev, int ht_cap);
 u8 pci_find_next_ht_capability(struct pci_dev *dev, u8 pos, int ht_cap);
-int pci_find_ext_capability(struct pci_dev *dev, int cap);
-int pci_find_next_ext_capability(struct pci_dev *dev, int pos, int cap);
+u16 pci_find_ext_capability(struct pci_dev *dev, int cap);
+u16 pci_find_next_ext_capability(struct pci_dev *dev, u16 pos, int cap);
 struct pci_bus *pci_find_next_bus(const struct pci_bus *from);
 
 u64 pci_get_dsn(struct pci_dev *dev);
